+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - CMS-1753 - Remove slug links from markdown fields
 
 ### Fixed Bugs
+- CMS-1830 - Move title field into translated admin panels to fix bug with untranslated tree_based_breadcrumbs
 
 
 ## [2019.08.15](https://github.com/uktrade/directory-cms/releases/tag/2019.08.15)